refactor(api): extract AIProvider seam over Workers AI (DEHARD-2)
Introduce a minimal AIProvider interface (generateTitle, generateFollowUps,
generateTitleAndFollowUps, generateEmbedding) with WorkersAIProvider as the
default implementation. The existing logic (models, prompts, parsing) is MOVED
verbatim from ai.ts into the provider; the exported service functions now
delegate to it, so all call sites keep passing `ai: Ai` unchanged.

Makes the AI provider swappable later (portability) without coupling callers to
Workers AI. Behavior is identical — same models (@cf/meta/llama-3.1-8b-instruct,
@cf/baai/bge-m3), same prompts, same parsing/fallbacks; vitest 280/280 unchanged.

Risk: 🟢 behavior-preserving extraction; no call-site or API change, no new deps.
Rollback: revert this PR.
